% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_model.R
\name{demo_model}
\alias{demo_model}
\title{Demo model}
\usage{
demo_model(T, params)
}
\arguments{
\item{T}{periods to simulate}

\item{params}{list of parameters}
}
\description{
Basic RBC model to use as a testing function. The model works through the package gEcon make sure this package is installed before running the function.
}
