% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{elim}
\alias{elim}
\title{elim}
\usage{
elim(D_bar, variance)
}
\arguments{
\item{D_bar}{vector of average distances for different CoPs}

\item{variance}{covariance matrix of D_bar}
}
\description{
elim
}
