% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{fast_ICA_model}
\alias{fast_ICA_model}
\title{fast_ICA_model}
\usage{
fast_ICA_model(residuals, n.comp, ...)
}
\arguments{
\item{residuals}{matrix of VAR residuals}

\item{n.comp}{number of components to identify}

\item{...}{see other fastICA comments}
}
\description{
Indentification of the model mixing matrix through Indepdent component analysis alogrithm "fastICA"
}
