% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{fastICA_gen}
\alias{fastICA_gen}
\title{fast_ICA_general}
\usage{
fastICA_gen(ures, sseed = 46, maxit = 3000, tol = 1e-14, verbose = FALSE)
}
\arguments{
\item{ures}{residuals of VAR model}

\item{sseed}{seed}

\item{maxit}{see fastICA docu}

\item{tol}{see fastICA docu}

\item{verbose}{see fastICA docu}
}
\description{
Indentification of the mixing matrix through Indepdent component analysis alogrithm "fastICA" and ordering of matrix
}
