% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{fastICA_model}
\alias{fastICA_model}
\title{fast_ICA_model}
\usage{
fastICA_model(
  residuals,
  sseed = 46,
  maxit = 3000,
  tol = 1e-14,
  verbose = FALSE
)
}
\arguments{
\item{residuals}{matrix of VAR residuals}

\item{sseed}{seed}

\item{maxit}{see fastICA docu}

\item{tol}{see fastICA docu}

\item{verbose}{see fastICA docu}
}
\description{
Indentification of the list of list model mixing matrix through Indepdent component analysis alogrithm "fastICA"
}
