% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{lag_MA_components}
\alias{lag_MA_components}
\title{lag_MA_components}
\usage{
lag_MA_components(var_est, Mixmat, maxlag = 5)
}
\arguments{
\item{var_est}{result of estimated VAR model}

\item{Mixmat}{estimated mixing matrix of contemporanous effects}

\item{maxlag}{nubmer of lagged coefficient to compute}
}
\description{
Function to extract the coefficients of  lagged componenets of the moving average represenetation of the VAR model
}
