% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{lag_MA_components_mod}
\alias{lag_MA_components_mod}
\title{lag_MA_components_mod}
\usage{
lag_MA_components_mod(var_est_model, Mixmat_model, maxlag = 5)
}
\arguments{
\item{var_est_model}{var estimates of the model in a list of lists}

\item{Mixmat_model}{estimates of the model mixing matrx in a list of lists}

\item{maxlag}{number of lagged components to compute}
}
\description{
Function to apply the lag_MA_components function to a list of list of model estimates
}
