% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{MCS}
\alias{MCS}
\title{MCS function}
\usage{
MCS(MDI_matrix, fun_p = p_values, fun_el = elim, dA = dA)
}
\arguments{
\item{MDI_matrix}{matrix of MDIs}

\item{fun_p}{function to determine p-value}

\item{fun_el}{elimination rule}

\item{dA}{significance threshold}
}
\value{
Model confidence set of one lagged component.
}
\description{
Function to determine the model confidence set and p-values
}
