% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{MCS_list}
\alias{MCS_list}
\title{MSC_list}
\usage{
MCS_list(MDI_matrix_list, dA = 1)
}
\arguments{
\item{MDI_matrix_list}{list of MDI matrices}

\item{dA}{significance level default dA=1 selet best cop}
}
\description{
Computes the minimal confidence set of list of MDI matrices
}
