% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_model.R
\name{model}
\alias{model}
\title{Demo model function of economic system}
\usage{
model(alpha, beta, gamma, delta, sigma, t, y_0, i_0)
}
\arguments{
\item{i_0}{}
}
\description{
Demo model function of economic system
}
