% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{p_values}
\alias{p_values}
\title{p_values}
\usage{
p_values(iN, dA, D_bar, variance)
}
\arguments{
\item{iN}{number of runs per CoP}

\item{dA}{significance theshold}

\item{D_bar}{vector of average distances for different CoPs}

\item{variance}{covariance matrix of D_bar}
}
\description{
Function to determine the p-values of minimal distance index based on Chi-squared
}
