% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{signperm_matrix}
\alias{signperm_matrix}
\title{Sign permutation function}
\usage{
signperm_matrix(A, B)
}
\arguments{
\item{A}{n by n matrix such that ||AP-B||f.}

\item{B}{n by n matrix such that ||AP-B||f.}
}
\value{
Sign permutation matrix that minimizes the frobenius norm between A and B
}
\description{
Compute optimal sign permutation matrix using linear assignment algorithm.
}
