% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_2.R
\name{VAR_emp}
\alias{VAR_emp}
\title{VAR__emp}
\usage{
VAR_emp(data, var_select, lag_select = NULL, ...)
}
\arguments{
\item{data}{Empirical data to be used.}

\item{var_select}{Variables to be compared to the model SVAR.}

\item{lag_select}{Integer for the lag order. The default is NULL in which case the optimum lag is determined based on the information criterion.}

\item{...}{
  Arguments passed on to \code{\link[vars:VARselect]{vars::VARselect}}, \code{\link[vars:VAR]{vars::VAR}}
  \describe{
    \item{\code{lag.max}}{Integer for the highest lag order (default is
    \code{lag.max = 10}).}
    \item{\code{type}}{Type of deterministic regressors to include.}
    \item{\code{season}}{Inlusion of centered seasonal dummy variables (integer
    value of frequency).}
    \item{\code{exogen}}{Inlusion of exogenous variables.}
    \item{\code{ic}}{Character, selects the information criteria, if
    \code{lag.max} is not \code{NULL}.}
  }}
}
\value{
A var estimation object.
}
\description{
VAR estimation based on empirical data
}
