% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_2.R
\name{VAR_model}
\alias{VAR_model}
\title{VAR_model}
\usage{
VAR_model(
  var_select,
  model_data,
  trim,
  lag_select,
  season = NULL,
  exog = NULL,
  type = "const"
)
}
\arguments{
\item{var_select}{select the variables to be compared to the real world SVAR.}

\item{model_data}{list of list of dataframes generated by the model.}

\item{trim}{the number of observation to trim from the data set from 0 - trim}

\item{lag_select}{Integer for the lag order (default is p=1).}

\item{season}{Inlusion of centered seasonal dummy variables (integer value of frequency).}

\item{exog}{Inlusion of exogenous variables.}

\item{type}{Type of deterministic regressors to include.}
}
\description{
VAR estimation based on model data
}
