% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{lag_MA_components}
\alias{lag_MA_components}
\title{Moving average components function}
\usage{
lag_MA_components(var_est, Mixmat, maxlag = 5)
}
\arguments{
\item{var_est}{Result of estimated VAR model.}

\item{Mixmat}{Estimated mixing matrix of contemporaneous effects.}

\item{maxlag}{Number of lagged coefficients to compute, default is set to 5.}
}
\value{
List of laged component matrices.
}
\description{
Function to extract the coefficients of lagged components of the moving average representation of the VAR model.
}
