% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{lag_MA_components_mod}
\alias{lag_MA_components_mod}
\title{lag_MA_components_mod}
\usage{
lag_MA_components_mod(var_est_model, Mixmat_model, ...)
}
\arguments{
\item{var_est_model}{List of list of VAR estimates of the model.}

\item{Mixmat_model}{List of lists of estimates of the model mixing matrix.}

\item{...}{
  Arguments passed on to \code{\link[=lag_MA_components]{lag_MA_components}}
  \describe{
    \item{\code{maxlag}}{Number of lagged coefficients to compute, default is set to 5.}
  }}
}
\value{
List of lists of lists of lagged component matrices.
}
\description{
Function to apply the lag_MA_components function to a list of list of model estimates
}
