% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_4.R
\name{MCS_list}
\alias{MCS_list}
\title{MSC_list}
\usage{
MCS_list(MDI_matrix_list, ...)
}
\arguments{
\item{MDI_matrix_list}{list of MDI matrices}

\item{...}{
  Arguments passed on to \code{\link[=MCS]{MCS}}
  \describe{
    \item{\code{dA}}{significance threshold}
    \item{\code{fun_p}}{function to determine p-value}
    \item{\code{fun_el}}{elimination rule}
  }}
}
\value{
List of lagged components consisting of the model confidence sets.
}
\description{
Computes the minimal confidence set of list of MDI matrices
}
