% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_1.R
\name{model_data}
\alias{model_data}
\title{Model Data Function}
\usage{
model_data(model, param_space, samples, runs)
}
\arguments{
\item{model}{function representing the theoretical model to be calibrated}

\item{param_space}{the parameters to be calibrated and their boundaries}

\item{samples}{number of samples to draw from the parameter space}

\item{runs}{the number of runs per parameter sample}
}
\value{
a matrix containing the model data for the specified number of runs and samples
}
\description{
Function to generate the model data for multiple samples of the parameter space and multiple runs
}
