% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_3.R
\name{resid_model}
\alias{resid_model}
\title{resid_model}
\usage{
resid_model(var_est)
}
\arguments{
\item{var_est}{VAR estimate based on model data}
}
\description{
Calculate the residuals based on the VAR estimate of the model data
}
