% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_2.R
\name{VAR_emp}
\alias{VAR_emp}
\title{VAR__emp}
\usage{
VAR_emp(
  data,
  var_select,
  lag_select,
  season = NULL,
  exog = NULL,
  type = "const"
)
}
\arguments{
\item{data}{Empirical data to be used}

\item{var_select}{Variables to be compared to the model SVAR}

\item{lag_select}{Integer for the lag order (default is p=1).}

\item{season}{Inlusion of centered seasonal dummy variables (integer value of frequency).}

\item{exog}{Inlusion of exogenous variables.}

\item{type}{Type of deterministic regressors to include.}
}
\description{
VAR estimation based on empirical data
}
