% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_2.R
\name{VAR_model}
\alias{VAR_model}
\title{VAR of model data}
\usage{
VAR_model(var_select, model_data, trim, lag_select = NULL, ...)
}
\arguments{
\item{var_select}{Select the variables to be compared to the real world SVAR.}

\item{model_data}{List of list of dataframes generated by the model.}

\item{trim}{The number of observation to trim from the data set from 0 - trim.}

\item{lag_select}{Integer for the lag order. The default is NULL in which case the optimum lag is determined based on the AIC.}

\item{...}{
  Arguments passed on to \code{\link[vars:VARselect]{vars::VARselect}}, \code{\link[vars:VAR]{vars::VAR}}
  \describe{
    \item{\code{lag.max}}{Integer for the highest lag order (default is
    \code{lag.max = 10}).}
    \item{\code{type}}{Type of deterministic regressors to include.}
    \item{\code{season}}{Inlusion of centered seasonal dummy variables (integer
    value of frequency).}
    \item{\code{exogen}}{Inlusion of exogenous variables.}
    \item{\code{ic}}{Character, selects the information criteria, if
    \code{lag.max} is not \code{NULL}.}
  }}
}
\value{
A list of list of VAR estimations.
}
\description{
VAR estimation where the input is a list of list of dataframes, where each dataframe represents a specific combination of a parameter setting and a simulation run.
}
